% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate}
\alias{animate}
\title{Render Data with D3}
\usage{
animate(data = function() floor(runif(6, 1, 40)),
  script = system.file("samples/barchart-animation.js", package = "r2d3"),
  width = NULL, height = NULL, inject = "data", version = "5.0.0",
  interval = 1000)
}
\arguments{
\item{data}{A function that provides data to be passed to D3.js script.}

\item{script}{The 'JavaScript' file containing the D3.js script.}

\item{width}{The desired width of the widget.}

\item{height}{The desired height of the widget.}

\item{inject}{The variable name used to inject data into a D3 script.}

\item{version}{The D3 version to use.}

\item{interval}{The animation inverval in milliseconds.}
}
\description{
Renders a data with D3 as an HtmlWidget using a generic D3.js script.
}
