% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_d3_data}
\alias{as_d3_data}
\alias{as_d3_data.default}
\title{Convert object to D3 data}
\usage{
as_d3_data(x, ...)

\method{as_d3_data}{default}(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{Additional arguments for generic methods}
}
\description{
Generic method to transform R objects into D3 friendly data.
}
\details{
The value returned from \code{as_d3_data()} should be one of:
\itemize{
\item An R data frame. In this case the \code{HTMLWidgets.dataframeToD3()}
JavaScript function will be called on the client to transform
the data into D3 friendly (row-oriented) data; or
\item A JSON object created using \link[jsonlite:toJSON]{jsonlite::toJSON}; or
\item Any other R object which can be coverted to JSON using \link[jsonlite:toJSON]{jsonlite::toJSON}.
}
}
