% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{d3_animate}
\alias{d3_animate}
\title{Render Data with D3}
\usage{
d3_animate(data = function() {     Sys.sleep(1)     floor(runif(6, 1, 40)) },
  script = system.file("samples/barchart-variable.js", package = "d3"),
  width = NULL, height = NULL, version = "5.0.0")
}
\arguments{
\item{data}{A function that provides data to be passed to D3.js script.}

\item{script}{The 'JavaScript' file containing the D3.js script.}

\item{width}{The desired width of the widget.}

\item{height}{The desired height of the widget.}

\item{version}{version The D3 version to use.}
}
\description{
Renders a data with D3 as an HtmlWidget using a generic D3.js script.
}
