% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{d3_available_extensions}
\alias{d3_available_extensions}
\title{Available D3 Extensions}
\usage{
d3_available_extensions(name = NULL, latest = TRUE)
}
\arguments{
\item{name}{The name of the extension. Defaults to \code{NULL} to
enumerate all available extensions.}

\item{latest}{Attempt to retrieve the latest extensions from repo?}
}
\description{
Provides a data frame of available D3 extensions that can be
used as input to \code{d3_extension()}.
}
\seealso{
\code{\link[=d3_extension]{d3_extension()}}
}
