% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{d3_dependency}
\alias{d3_dependency}
\title{D3 HTML dependency}
\usage{
d3_dependency(version = c("5", "4", "3"))
}
\arguments{
\item{version}{Major version of D3}
}
\description{
Create an HTML dependency for the D3 library.
}
\details{
Create an HTML dependency for a version of D3. Each version has
a distinct root D3 object so it's possible to combine multiple versions
of D3 on a single page. For example, D3 v5 is accessed via \code{d3v5} and
D3 v4 is accessed via \code{d3v4}. Note however that D3 v3 is accessed via
simply \code{d3} (for compabibilty with existing htmlwidgets that use
this form).
}
\note{
This function is exported for use by htmlwidgets. If you are
using the \code{r2d3()} function to include D3 code within a document
or application this dependency is included automatically so calling
this function is unnecessary.
}
