% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extensions.R
\name{d3_extension}
\alias{d3_extension}
\title{D3 Extension}
\usage{
d3_extension(name, target = tempdir())
}
\arguments{
\item{name}{The name of the D3 extension}

\item{target}{Target path to download the extension to, defaults to \code{tempdir()}.}
}
\description{
Create a R2D3 dependency from a D3 extension, a list of available
extensions can be retrieved using \code{d3_available_extensions()}.
}
\seealso{
\code{\link[=d3_available_extensions]{d3_available_extensions()}}
}
