% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{d3_render}
\alias{d3_render}
\title{Render Data with D3}
\usage{
d3_render(data = islands, script = system.file("samples/bubbles.js", package
  = "d3"), width = NULL, height = NULL, version = "5.0.0")
}
\arguments{
\item{data}{The data to be passed to D3.js script, currently only data frames supported.}

\item{script}{The 'JavaScript' file containing the D3.js script.}

\item{width}{The desired width of the widget.}

\item{height}{The desired height of the widget.}
}
\description{
Renders a data with D3 as an HtmlWidget using a generic D3.js script.
}
