% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D3.R
\name{D3Chord}
\alias{D3Chord}
\title{D3 Chord}
\usage{
D3Chord(JSON, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{file_out}{the location and name for the output html file}
}
\description{
Creates a html file containing json file and a D3.js Chord diagram
Needs json data from the jsonMatrix function
}
\examples{
dm<-matrix(c(11975,1951,8010,1013,5871,10048,16145,990,8916,2060,8090,940,2868,6171,8045,6907), 4,4)
labels<-c("Group1", "Group2", "Group3", "Group4")
json<-jsonMatrix(dm, labels)
D3Chord(json, file_out="chord.html")
}
\author{
James Thomson
}
\references{
https://bl.ocks.org/mbostock/4062006
}

