% Generated by roxygen2 (4.0.1): do not edit by hand
\name{D3Dendo}
\alias{D3Dendo}
\title{D3 Denogram}
\usage{
D3Dendo(JSON, text = 15, height = 800, width = 700, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{text}{the text size in the dendogram}

\item{height}{the height of the denogram}

\item{weidth}{the width of the dendogram}

\item{the}{location and name for the output html file}
}
\description{
Creates a html file containing json file and a D3.js Dendogram
}
\examples{
hc <- hclust(dist(USArrests), "ave")
plot(hc)
JSON<-HCtoJSON(hc)
D3Dendo(JSON, file_out="/Users/home/Documents/R_Projects/D3/USArrests_Dendo.html")
}
\author{
James Thomson
}

