% Generated by roxygen2 (4.0.1): do not edit by hand
\name{D3Dendro}
\alias{D3Dendro}
\title{D3 Dendrogram}
\usage{
D3Dendro(JSON, text = 15, height = 800, width = 700, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{text}{the text size in the dendrogram}

\item{height}{the height of the dendrogram}

\item{weidth}{the width of the dendrogram}

\item{the}{location and name for the output html file}
}
\description{
Creates a html file containing json file and a D3.js Dendrogram
}
\examples{
hc <- hclust(dist(USArrests), "ave")
plot(hc)
JSON<-HCtoJSON(hc)
D3Dendro(JSON, file_out="USArrests_Dendo.html")
}
\author{
James Thomson
}

