% Generated by roxygen2 (4.0.1): do not edit by hand
\name{D3Force}
\alias{D3Force}
\title{D3 Force}
\usage{
D3Force(JSON, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{the}{location and name for the output html file}
}
\description{
Creates a html file containing json file and a D3.js Force Directed Layout.
If you want toe colours in teh force directed layout to represent a group. PLease ensure the column is labelled "group" in the data frame
}
\examples{
nodes.df<-data.frame(name=c("Dan", "Digby", "Lex", "Flamer", "Stripey"), age=c(32, 38, 45, 17, 2))
links.df<-data.frame(source=c("Dan", "Digby", "Flamer"), target=c("Lex", "Flamer", "Stripey"))
JSON<-jsonNodesLinks(nodes.df, links.df)
D3Force(JSON, file_out="Force.html")
}
\author{
Simon Raper and James Thomson
}

