% Generated by roxygen2 (4.0.1): do not edit by hand
\name{D3Sankey}
\alias{D3Sankey}
\title{D3 Sankey}
\usage{
D3Sankey(JSON, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{the}{location and name for the output html file}
}
\description{
Creates a html file containing json file and a D3.js Sankey
}
\examples{
hc.ave <- hclust(dist(USArrests), "ave")
hc.single <- hclust(dist(USArrests), "single")
hc.ward <- hclust(dist(USArrests), "ward.D")

cut.ave<-cutree(hc.ave, k=6)
cut.single<-cutree(hc.single, k=6)
cut.ward<-cutree(hc.ward, k=6)

ClustComp<-data.frame(States=rownames(USArrests), ave=as.vector(cut.ave),single=as.vector(cut.single),ward=as.vector(cut.ward))

JSON<-jsonSankey(ClustComp)
D3Sankey(JSON, file_out="Sankey.html")
}
\author{
Simon Raper and James Thomson
}

