% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D3.R
\name{D3Sunburst}
\alias{D3Sunburst}
\title{D3 Sunburst}
\usage{
D3Sunburst(JSON, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{file_out}{the location and name for the output html file}
}
\description{
Creates a html file containing json file and a D3.js Sunburst.
Needs json data from the jsonSeqFreq function
}
\examples{
data<-read.csv(file="https://gist.githubusercontent.com/kerryrodden/7090426/raw/8fce22c9e21711c757ee8a0df7dba5a42dea0d9c/visit-sequences.csv",
header = FALSE,  stringsAsFactors = FALSE
)
seq<-data[-13632,1]
freq<-data[-13632,2]
json<-jsonSeqFreq(seq, freq)
D3Sunburst(json, file_out="sunburst.html")
}
\author{
James Thomson
}
\references{
https://bl.ocks.org/kerryrodden/7090426
}

