% Generated by roxygen2 (4.0.1): do not edit by hand
\name{D3Tree}
\alias{D3Tree}
\title{D3 Tree}
\usage{
D3Tree(JSON, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{the}{location and name for the output html file}
}
\description{
Creates a html file containing json file and a D3.js Tree Map.
The nested json needs values assigned to it in order for it to work
}
\examples{
data(counties)
JSON<-jsonNestedData(structure=counties[,1:3], values=counties[,4], top_label="UK")
D3Tree(JSON, file_out="Tree.html")
}
\author{
James Thomson
}

