% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D3.R
\name{D3Venn}
\alias{D3Venn}
\title{D3 Venn}
\usage{
D3Venn(JSON, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{the}{location and name for the output html file}
}
\description{
Creates a html file containing json file and a D3.js Venn diagram.
The nested json needs values assigned to it in order for it to work
}
\examples{
data(browsers)
JSON<-jsonOverlaps(browsers, overlaps = 4)
D3Venn(JSON, file_out="browsers_venn.html")
}
\author{
James Thomson
}
\references{
Ben Frederickson d3 Venn library: https://github.com/benfred/venn.js
}

