% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D3.R
\name{D3WordCloud}
\alias{D3WordCloud}
\title{D3 Word Cloud}
\usage{
D3WordCloud(JSON, file_out)
}
\arguments{
\item{JSON}{A json object}

\item{file_out}{the location and name for the output html file}
}
\description{
Creates a html file containing json data and a D3.js Word cloud.
It works with the jsonWordCloud function
}
\examples{
words=c("big", "data", "machine", "learning", "wordcloud", "R", "d3js", "algorithm", "analytics", "science", "API")
freq=c(50, 50, 30, 30, 100, 10, 10, 10, 5, 5, 5 )
json<-jsonwordcloud(words, freq)
D3WordCloud(json, file_out="word_cloud.html")
}
\author{
James Thomson
}
\references{
https://github.com/jasondavies/d3-cloud
}

