% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{default_sizing}
\alias{default_sizing}
\title{Default sizing policy for r2d3 widgets}
\usage{
default_sizing()
}
\description{
Default sizing policy for r2d3 widgets
}
\details{
Use \code{\link[htmlwidgets:sizingPolicy]{htmlwidgets::sizingPolicy()}} to specify an
alternate policy.
}
\keyword{internal}
