% Generated by roxygen2 (4.0.1): do not edit by hand
\name{HCtoJSON}
\alias{HCtoJSON}
\title{Hierachical Cluster to json}
\usage{
HCtoJSON(hc)
}
\arguments{
\item{hc}{A hieracichal cluster (hclust) output object}
}
\description{
Creates a json file from a hierachical clustering output
}
\examples{
hc <- hclust(dist(USArrests), "ave")
plot(hc)
JSON<-HCtoJSON(hc)
}
\author{
James Thomson
}

