% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jsonCompare}
\alias{jsonCompare}
\title{Multiple results to json}
\usage{
jsonCompare(data)
}
\arguments{
\item{data}{A data frame with first column the record identifier and a series of columns showig the allocations to compare}
}
\description{
Creates a json file from a sequence of events. This json will work with D3Sankey
}
\examples{
hc.ave <- hclust(dist(USArrests), "ave")
hc.single <- hclust(dist(USArrests), "single")
hc.ward <- hclust(dist(USArrests), "ward.D")

cut.ave<-cutree(hc.ave, k=6)
cut.single<-cutree(hc.single, k=6)
cut.ward<-cutree(hc.ward, k=6)

ClustComp<-data.frame(States=rownames(USArrests), ave=as.vector(cut.ave),single=as.vector(cut.single),ward=as.vector(cut.ward))

JSON<-jsonCompare(ClustComp)
}
\author{
Simon Raper and James Thomson
}

