% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.R
\name{jsonOverlaps}
\alias{jsonOverlaps}
\title{Data frame to overlaps json}
\usage{
jsonOverlaps(data, overlaps = 2)
}
\arguments{
\item{data}{A data frame to be converted to json. The first column should represent the item, and the second column should represent the group it belongs to}

\item{overlaps}{The degree of overlaps to be considered. Defaults to the total number of groups}
}
\description{
Creates a json file representing overlaps from a data frame. This json will work with D3 Venn
}
\examples{
data(browsers)
JSON<-jsonOverlaps(browsers, overlaps = 4)
}
\author{
James Thomson
}

