% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.R
\name{jsonSeqFreq}
\alias{jsonSeqFreq}
\title{Sequence frequency json}
\usage{
jsonSeqFreq(seq, freq)

jsonSeqFreq(seq, freq)
}
\arguments{
\item{seq}{A vector of sequences. Each item in sequence must be seperated by a -. For example item1-item2-item3}

\item{freq}{A vector of frequencies for the sequences}

\item{seq}{A vector of sequences. Each item in sequence must be seperated by a -. For example item1-item2-item3}

\item{freq}{A vector of frequencies for the sequences}
}
\description{
Creates a json file of sequence frequencies. This json will work with D3 Sunburst

Creates a json file of sequence frequencies. This json will work with D3 Sunburst
}
\examples{
data<-read.csv(file="https://gist.githubusercontent.com/kerryrodden/7090426/raw/8fce22c9e21711c757ee8a0df7dba5a42dea0d9c/visit-sequences.csv",
header = FALSE,  stringsAsFactors = FALSE
)
seq<-data[-13632,1]
freq<-data[-13632,2]
jsonSeqFreq(seq, freq)
data<-read.csv(file="https://gist.githubusercontent.com/kerryrodden/7090426/raw/8fce22c9e21711c757ee8a0df7dba5a42dea0d9c/visit-sequences.csv",
header = FALSE,  stringsAsFactors = FALSE
)
seq<-data[-13632,1]
freq<-data[-13632,2]
jsonSeqFreq(seq, freq)
}
\author{
James Thomson

James Thomson
}

