% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.R
\name{jsonWordCloud}
\alias{jsonWordCloud}
\title{Data to word frequency json}
\usage{
jsonWordCloud(words, freq = NULL)
}
\arguments{
\item{words}{A list of the words. Either a full list. Or a unique list and use the freq argument}

\item{freq}{Optional a list of frequencies for the words}
}
\description{
Creates a json file of word frequencies. This json will work with D3 Word Cloud
}
\examples{
words=c("big", "data", "machine", "learning", "wordcloud", "R", "d3js", "algorithm", "analytics", "science", "API")
freq=c(50, 50, 30, 30, 100, 10, 10, 10, 5, 5, 5 )
json<-jsonWordCloud(words, freq)
}
\author{
James Thomson
}

