% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/json.R
\name{jsonXtabs}
\alias{jsonXtabs}
\title{Data frame to cross tabs json}
\usage{
jsonXtabs(data)
}
\arguments{
\item{data}{A data frame to be converted to json. It should contain two columns one for each categorical variable}
}
\description{
Creates a json file representing between two categorical variables. This json will work with D3 Heat
}
\examples{
data<-data.frame(airquality$Month, airquality$Temp)
json<-jsonXtabs(data)
}
\author{
James Thomson
}

