% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{r2d3}
\alias{r2d3}
\title{Render Data with D3}
\usage{
r2d3(data, script, options = NULL, tag = "svg", version = c(4, 5),
  dependencies = NULL)
}
\arguments{
\item{data}{The data to be passed to D3 script.}

\item{script}{The 'JavaScript' file containing the D3 script.}

\item{options}{The options to be passed to D3 script.}

\item{tag}{The html tag to create for the D3 script.}

\item{version}{The major D3 version to use, the latest minor version
is automatically picked.}

\item{dependencies}{Additional javascript or css dependencies.}
}
\description{
Renders data with D3 as an HtmlWidget using a D3.js script.
}
