% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{sizingPolicy}
\alias{reexports}
\alias{read_json}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{htmlwidgets}{\code{\link[htmlwidgets]{sizingPolicy}}}

  \item{jsonlite}{\code{\link[jsonlite]{read_json}}}
}}

