% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/D3.R
\name{D3Force}
\alias{D3Force}
\title{D3 Force}
\usage{
D3Force(JSON, file_out, arrows = FALSE, collision.detect = FALSE,
  fisheye = FALSE)
}
\arguments{
\item{JSON}{A json object}

\item{the}{location and name for the output html file}

\item{options}{A list of features to include the graph (see the details section)}
}
\description{
Creates a html file containing json file and a D3.js Force Directed Layout.
If you want toe colours in teh force directed layout to represent a group. PLease ensure the column is labelled "group" in the data frame
}
\examples{
nodes.df<-data.frame(name=c("Dan", "Digby", "Lex", "Flamer", "Stripey"), age=c(32, 38, 45, 17, 2))
links.df<-data.frame(source=c("Dan", "Digby", "Flamer"), target=c("Lex", "Flamer", "Stripey"))
JSON<-jsonNodesLinks(nodes.df, links.df)
D3Force(JSON, file_out="Force.html")

#With directional arrows
D3Force(JSON, file_out="Force.html", options=list(arrows=TRUE))

data(celebs)
colnames(celebs$relationships)<-c('source', 'target')
colnames(celebs$celebs)<-c('name', 'group')
JSON<-jsonNodesLinks(celebs$celebs, celebs$relationships)
D3Force(JSON, file_out="/Users/home/Documents/R_Projects/Force.html")
}
\author{
Simon Raper and James Thomson
}
\references{
Mike Bostock's lovely d3: http://d3js.org/
}

