% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfToJSON}
\alias{dfToJSON}
\title{Converts data to json form}
\usage{
dfToJSON(df, mode = "vector")
}
\arguments{
\item{df}{a data.frame to be converted}

\item{mode}{there are three modes "vector", "coords" , "rowToObject"}
}
\description{
Creates a json file from an input
}
\examples{
example
}
\author{
Simon Raper
}

