% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dfToJSON}
\alias{dfToJSON}
\title{Multiple results to json}
\usage{
dfToJSON(df, mode = "vector")
}
\arguments{
\item{df}{a data.frame to be converted}

\item{mode}{there are two modes vector  and row to object}
}
\description{
Creates a json file from a sequence of events. This json will work with D3Sankey
}
\examples{
example
}
\author{
Simon Raper
}

