% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jsonHC}
\alias{jsonHC}
\title{Hierachical Cluster output to json}
\usage{
jsonHC(hc)
}
\arguments{
\item{hc}{A hieracichal cluster (hclust) output object}
}
\description{
Creates a json file from a hierachical clustering output. This json will work with D3 Dendro
}
\examples{
hc <- hclust(dist(USArrests), "ave")
plot(hc)
JSON<-jsonHC(hc)
}
\author{
James Thomson
}

