% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{r2d3}
\alias{r2d3}
\title{Render Data with D3}
\usage{
r2d3(data = floor(runif(6, 1, 40)),
  script = system.file("samples/barchart-variable.js", package = "r2d3"),
  width = NULL, height = NULL, dependencies = NULL, options = NULL,
  version = "5.0.0")
}
\arguments{
\item{data}{The data to be passed to D3.js script.}

\item{script}{Script(s) to include within the document.}

\item{width}{The desired width of the widget.}

\item{height}{The desired height of the widget.}

\item{version}{The D3 version to use.}

\item{d3}{The 'JavaScript' file containing the D3.js script.}

\item{inject}{The variable name used to inject data into a D3 script.}

\item{stylesheet}{Stylesheet(s) to include within the document.}
}
\description{
Renders data with D3 as an HtmlWidget using a D3.js script.
}
