% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{r2d3}
\alias{r2d3}
\title{Render Data with D3}
\usage{
r2d3(data, script, options = NULL, container = "svg", version = c(5, 4,
  3), dependencies = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{Data to be passed to D3 script.}

\item{script}{'JavaScript' file containing the D3 script.}

\item{options}{Options to be passed to D3 script.}

\item{container}{The 'HTML' container of the D3 output.}

\item{version}{Major D3 version to use, the latest minor version
is automatically picked.}

\item{dependencies}{Additional javascript or css dependencies.}

\item{width}{Desired width for output widget.}

\item{height}{Desired height for output widget.}
}
\description{
Renders data with D3 as an HtmlWidget using a D3.js script.
}
